Add model class `Robot`
This also adds a management command `add_robot` for creating new Robot instances that are empty except for the manufacturer ID.
For the future, it is intended that new Robot instances can also be added via the Django Admin.

It is intended that in the future the details of these Robot instances are updated by the management command `query_Husqvarna` or appropriate views with similar functionality.